Add dcline functionality. dclinecost not added yet.
This commit adds the dcline functionality to most.
The mpc file must be converted to internal numbering (ext2int)
before most is called. The output from most must be converted
back to external numbering in order to read the output in
mdo.flow(t).mpc.dcline.
If not converted, the dcline output is still attached to
the mdo.flow(t).mpc.gen data.