Extract values from dummy generator and write to dcline field.
This change calls run_userfcn with int2ext argument to extract
data from dummy generator. The dummy generators are removed.
This process is applied to mdo.flow as well as mdo.mpc.
Since ng is number of regular generators + number of dummy
generators, the call has to be made at this location in the code
to avoild trubble after removing the generators.